refactor: api 폴더 생성 후 구조 변경
